Error handling.

+++ icanc/commands/create.py
@@ -3,7 +3,8 @@ import os
 import shutil
 import subprocess
 import tomllib
-from .common.paths import icanc_path
+from .common.exception import FoundException, NotFoundException
+from .common.paths import ensure_paths, icanc_path
 
 @click.group()
 def create():
@@ -13,23 +14,34 @@ def create():
 @create.command()
 @click.argument("judge", type=str)
 @click.argument("problem", type=str)
-@click.option("--template", type=click.Path(exists=True), required=True)
+@click.option("--template", type=str, required=True)
 @click.option("--solution", "solution_dst", default="solution", help="Name for the solution file.")
 @click.option("--open", "open_editor", is_flag=True, help="Open solution file on text editor.")
 def solution(**kwargs):
     create_solution(**kwargs)
 
 def create_solution(judge, problem, template, solution_dst, open_editor):
+    ensure_paths()
+
+    template_filename = f"{template}.c"
+    template_path = icanc_path("templates", template_filename)
+    if not os.path.exists(template_path):
+        raise NotFoundException("template", f"./templates/{template_filename}")
+    
     with open(os.path.join(os.getcwd(), "icancrc.toml"), "rb") as f:
         cfg = tomllib.load(f)
     
     dir = icanc_path("problems", judge, problem)
     os.makedirs(dir, exist_ok=True)
 
-    solution_path = icanc_path("problems", judge, problem, f"{solution_dst}.c")
-    shutil.copy2(template, solution_path)
+    solution_filename = f"{solution_dst}.c"
+    solution_path_rel = f"./problems/{judge}/{problem}/{solution_filename}"
+    solution_path = icanc_path("problems", judge, problem, solution_filename)
+    if os.path.exists(solution_path):
+        raise FoundException("solution", solution_path_rel, "To create multiple solutions set the solution name with the --solution option.")
+    shutil.copy2(template_path, solution_path)
     
-    click.echo("Created blank solution: ./problems/{}/{}/{}.c".format(judge, problem, solution_dst))
+    click.echo(f"Created blank solution: {solution_path_rel}")
 
     if open_editor:
         subprocess.run([cfg["editor"], solution_path])
@@ -48,11 +60,15 @@ def create_testcases(judge, problem, testcases_dst, open_editor):
     
     dir = icanc_path("problems", judge, problem)
     if not os.path.exists(dir):
-        click.echo("Problem {}/{} does not exist.".format(judge, problem), err=True)
-        exit(1)
+        raise NotFoundException("problem", f"{judge}/{problem}")
     os.makedirs(dir, exist_ok=True)
-    testcases_path = icanc_path("problems", judge, problem, f"{testcases_dst}.toml")
-    
+
+    testcases_filename = f"{testcases_dst}.toml"
+    testcases_path_rel = f"./problems/{judge}/{problem}/{testcases_filename}"
+    testcases_path = icanc_path("problems", judge, problem, testcases_filename)
+    if os.path.exists(testcases_path):
+        raise FoundException("testcases", testcases_path_rel, "To create multiple testcases set the solution name with the --testcases option.")
+
     with open(testcases_path, "w") as f:
         f.writelines([
             "# {}/{}\n\n".format(judge, problem),
@@ -61,7 +77,7 @@ def create_testcases(judge, problem, testcases_dst, open_editor):
             "out = \"\"\"\"\"\"\n"
         ])
 
-    click.echo("Created blank testcases: ./problems/{}/{}/{}.toml".format(judge, problem, testcases_dst))
+    click.echo(f"Created blank testcases: {testcases_path_rel}")
 
     if open_editor:
         subprocess.run([cfg["editor"], testcases_path])
@@ -75,8 +91,7 @@ def create_testcases(judge, problem, testcases_dst, open_editor):
 @click.option("--open", "open_editor", is_flag=True, help="Open solution file on text editor.")
 def scaffold(judge, problem, template, solution_dst, testcases_dst, open_editor):
     """Create a solution and testcase files."""
-
-    click.echo("Scaffolding {}/{}\n".format(judge, problem))
-
+    
+    click.echo("Scaffolding {}/{}\n.".format(judge, problem))
     create_solution(judge, problem, template, solution_dst, open_editor)
     create_testcases(judge, problem, testcases_dst, False)
